% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian_copula.R
\name{fit_gaussian_copula}
\alias{fit_gaussian_copula}
\title{Fit Bivariate Gaussian Copula Model}
\usage{
fit_gaussian_copula(
  U,
  iter = 2000,
  chains = 4,
  warmup = 1000,
  thin = 1,
  seed = NULL,
  control = list(adapt_delta = 0.95, max_treedepth = 10),
  cores = 1
)
}
\arguments{
\item{U}{Data matrix of uniform marginals (U(0,1)), with two variables.}

\item{iter}{Number of iterations for each chain. Default is 2000.}

\item{chains}{Number of chains. Default is 4.}

\item{warmup}{Number of warmup iterations per chain. Default is 1000.}

\item{thin}{Thinning rate. Default is 1.}

\item{seed}{Random seed. Default is NULL.}

\item{control}{A list of parameters to control the sampler's behavior. Default is list(adapt_delta = 0.95, max_treedepth = 10).}

\item{cores}{Number of cores to use for parallel processing. Default is 1.}
}
\value{
A fitted Stan model and extracted correlation matrix.
}
\description{
Fit Bivariate Gaussian Copula Model
}
