% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Get the compiled Stan model (compiles on first use, caches for reuse)}
\usage{
get_stan_model()
}
\description{
Get the compiled Stan model (compiles on first use, caches for reuse)
}
\keyword{internal}
